requirements.txt
NumPy >= 1.11.1
Pytorch >= 0.3.0
nltk
tensorboard-pytorch(https://github.com/lanpa/tensorboard-pytorch) (build from source)